requirements.txt
Django
djangorestframework
drf-yasg
# drf-nested-routers # Entfernt
gunicorn
dj-database-url
whitenoise
# Neu für CORS:
django-cors-headers
# WebSocket Support:
channels
channels-redis
daphne
mysqlclient
